Changed Hero and Monster ACD value type

+++ Hero.cpp
@@ -115,7 +115,7 @@ Hero Hero::parse(const std::string& name) {
 			values.get<std::string>("name"),
 			values.get<int>("base_health_points"),
 			values.get<int>("base_damage"),
-			values.get<float>("base_attack_cooldown"),
+			values.get<double>("base_attack_cooldown"),
 			
 			values.get<int>("experience_per_level"),
 			values.get<int>("health_point_bonus_per_level"),
